Fold EditDetails CommonPageTests into the page test classes
Journeys don't get a CommonPageTests class; each page's tests live with
that page. The theories parameterised by page are split across
PersonalDetailsTests, NameChangeReasonTests, OtherDetailsChangeReasonTests
and CheckAnswersTests, each keeping only its own rows. Where the body
branched on the page name to pick form content, the branch is now
resolved statically. Same 241 tests as before.
